SW-2076 Add PageHeaderWrapper to Pages
- Added the header wrapper to several more pages
- species list
- accessions list
- check-in accessions
- monitoring
- inventory list
